feature/permissions-on-metadata-service: Refactor AuthorisedToEditMetadataOrUploader to make it more general
