+++ media-api/app/controllers/MediaApi.scala
@@ -56,6 +56,11 @@ class MediaApi(
 
   private val searchLink = Link("search", searchLinkHref)
 
+
+  private def getUploader(imageId: String, user: Principal): Future[Option[String]] = elasticSearch.getImageUploaderById(imageId)
+
+  private def authorisedForDeleteImageOrUploader(imageId: String) = authorisation.actionFilterForUploaderOr(imageId, DeleteImage, getUploader)
+
   private def indexResponse(user: Principal) = {
     val indexData = Json.obj(
       "description" -> "This is the Media API"
