+++ metadata-editor/app/controllers/EditsController.scala
@@ -7,7 +7,7 @@ import com.amazonaws.AmazonServiceException
 import com.gu.mediaservice.GridClient
 import com.gu.mediaservice.lib.argo.ArgoHelpers
 import com.gu.mediaservice.lib.argo.model._
-import com.gu.mediaservice.lib.auth.Authentication.Request
+import com.gu.mediaservice.lib.auth.Authentication.{Principal, Request}
 import com.gu.mediaservice.lib.auth.Permissions.EditMetadata
 import com.gu.mediaservice.lib.auth.{Authentication, Authorisation, SimplePermission}
 import com.gu.mediaservice.lib.aws.{NoItemFound, UpdateMessage}
@@ -59,28 +59,11 @@ class EditsController(
   val services: Services = new Services(config.domainRoot, ServiceHosts.guardianPrefixes, Set.empty)
   val gridClient: GridClient = GridClient(services)(ws)
 
-  def AuthorisedToEditMetadataOrUploader(imageId: String): ActionFilter[Request] = new ActionFilter[Request] {
-    override protected def filter[A](request: Request[A]): Future[Option[Result]] = {
-      //We first check for permissions, if the user has permissions we avoid making a call to media-api service
-      val hasPermission: Boolean = authorisation.hasPermissionTo(EditMetadata)(request.user)
-      if (hasPermission) {
-        Future.successful(None)
-      } else {
-        val result = for {
-          uploadedBy <- gridClient.getUploadedBy(imageId, auth.getOnBehalfOfPrincipal(request.user))
-          isAuthorised = authorisation.canUserWriteMetadata(request.user, uploadedBy.getOrElse(""))
-          if isAuthorised
-        } yield {
-          Future.successful(None)
-        }
-        result.flatten.recover{case _ => Some(authorisation.unauthorized(EditMetadata))}
-      }
-    }
-    override protected def executionContext: ExecutionContext = ec
-  }
-
-
   val metadataBaseUri = config.services.metadataBaseUri
+
+  private def getUploader(imageId: String, user: Principal): Future[Option[String]] = gridClient.getUploadedBy(imageId, auth.getOnBehalfOfPrincipal(user))
+
+  private def authorisedForEditMetadataOrUploader(imageId: String) = authorisation.actionFilterForUploaderOr(imageId, EditMetadata, getUploader)
 
   def decodeUriParam(param: String): String = decode(param, "UTF-8")
 
@@ -194,7 +177,7 @@ class EditsController(
     }
   }
 
-  def setMetadata(id: String) = (auth andThen AuthorisedToEditMetadataOrUploader(id)).async(parse.json) { req =>
+  def setMetadata(id: String) = (auth andThen authorisedForEditMetadataOrUploader(id)).async(parse.json) { req =>
     (req.body \ "data").validate[ImageMetadata].fold(
       errors => Future.successful(BadRequest(errors.toString())),
       metadata =>
@@ -204,7 +187,7 @@ class EditsController(
     )
   }
 
-  def setMetadataFromUsageRights(id: String) = (auth andThen AuthorisedToEditMetadataOrUploader(id)).async { req =>
+  def setMetadataFromUsageRights(id: String) = (auth andThen authorisedForEditMetadataOrUploader(id)).async { req =>
     store.get(id) flatMap { dynamoEntry =>
       val edits = dynamoEntry.as[Edits]
       val originalMetadata = edits.metadata
